# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/languages": {
            "get": {
                "description": "Возвращает все доступные языки программирования с опциональной фильтрацией по названию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Языки программирования"
                ],
                "summary": "Получить все языки программирования",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название для фильтрации",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.GetAllLanguagesResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавить новый язык программирования",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Языки программирования"
                ],
                "summary": "Добавить язык программирования",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Изображение",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Название",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Предмет",
                        "name": "subject",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Задание",
                        "name": "task",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Описание задания",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/languages/{id}": {
            "get": {
                "description": "Возвращает более подробную информацию об одном языке программирования",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Языки программирования"
                ],
                "summary": "Получить один язык программирования",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id языка программирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Language"
                        }
                    }
                }
            },
            "put": {
                "description": "Изменить данные полей об языке программирования",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Языки программирования"
                ],
                "summary": "Изменить язык программирования",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор языка программирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Предмет",
                        "name": "subject",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Задание",
                        "name": "task",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Изображение",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Описание задания",
                        "name": "description",
                        "in": "formData"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Удаляет язык программирования по id",
                "tags": [
                    "Языки программирования"
                ],
                "summary": "Удалить язык программирования",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id языка программирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/languages/{id}/add_to_form": {
            "post": {
                "description": "Добавить выбранный язык программирования в черновик формы",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Языки программирования"
                ],
                "summary": "Добавить в форму",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id языка программирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemes.AddToFormResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.Language": {
            "type": "object",
            "required": [
                "description",
                "name",
                "subject",
                "task"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "task": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schemes.AddToFormResp": {
            "type": "object",
            "properties": {
                "language_count": {
                    "type": "integer"
                }
            }
        },
...
        "schemes.GetAllLanguagesResponse": {
            "type": "object",
            "properties": {
                "draft_form": {
                    "type": "string"
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Language"
                    }
                }
            }
        },
...
    }
}